Python/requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
idna==3.7
pillow==10.3.0
PyYAML==6.0.1
requests==2.32.3
urllib3==2.2.1
